requirements.txt
aiofiles
Flask
google_cloud_aiplatform
numpy
pandas
protobuf
python-dotenv
Requests
setuptools
setuptools
streamlit
streamlit-webrtc
